Fix missing ADPCMA samples
When a new aon event is received, if the last sample played ends with same address as the new one, a done is immediately triggered. This means if the same sample is played repeatedly, every other one will be dropped. The same can happen for other samples with same end address, even if they are in different banks. This fixes the issue.
This also uncomments jotego's code for the end address to use the final samples (i.e. address ends in 0xFF), which appears to match the documentation.