Decode looks like it's only 12 bits and wraps. (#47)
* Decode looks like it's only 12 bits and wraps.

There are still a few changes in the experimental build of MiSTer Neo Geo that are different from your base version.  I'd like to get them back in sync.  I am less certain these are correct.  This is the one that looks most likely correct to me.  Please take a look and comment.
This change is based on this:
https://github.com/mamedev/mame/blob/master/src/devices/sound/fm.cpp#L2502

I believe this means the value is limited to 12 bits, but wraps, instead of clipping.  It looks like a relatively recent intentional change to the code base, which I assume is based on recent findings.  As I recall this fixed the sound in something, but I forget where (Metal Slug, maybe?).

It might be better to return this as a 12 bit signed value, rather than a 16 bit signed value (where the top 4 bits are always just sign extended).

This will change audio balance in other places, which are handled in other changes in the MiSTer core.  Not sure if you want to look at all the remaining changes at once, or one at a time.

* Don't clip LUT values

Not sure if this is correct or not.  This allows wrapping of the accumulator values, rather than clipping at o3777 (11 bits).  This is required if the ADPCMA decoder wraps at 12 bits rather than clipping. i.e. If either the accumulator or the decoder wraps, both should.

* Don't skip first sample

The first sample (when addr1 == {start, 9'd0}) is not being processed because clr is also set.  There's probably a better way to code this.