Bank only uses start address, not end address
Lots of documentation states that start and end address need to be in the same bank. However, I haven't read anything that says what happens when this is done against recommendation.  Just found out that Twinkle Star Sprites select screen does this. It sets bank = 3 for the start address, then sets end address to bank 2.  It then expects to play the sample in bank 3.  I take this to mean that only start updates the bank, and the high nibble for end address is never used.  This change fixes the Twinkle Star Sprites song (FC1E).