Current/Enable Channel Timing change needed for MiSTer
This change may not actually be a change to match hardware behavior, but is needed to work with MiSTer.  If cur_ch and en_ch are shifted the same direction {(0,0), (1,1), (2,2), (3,3), (4,4), (5,5)}, then the time between 0,0 and 1,1 is 7 cycles, and the time between (5,5) and (0,0) is 1 cycle (7 * 5 + 1 * 1 = 36 total cycles).  MiSTer does not like the 1 cycle difference when reading the sample ROMs.  This change inverts the en_ch variable so the timing is instead (5 * 5 + 11 * 1 = 36 total cycles).  This gives a minimum time between samples of 5 cycles.  I tried just moving the en_ch instead of inverting the shift order, and that just moved which channel had the 1 cycle minimum (and which channel malfunctioned because of bad sample ROM reading).  If this doesn't match real hardware, maybe a parameter (accurate vs slow timing requirement) could be added for implementations like MiSTer that require larger minimum read spacing?