Change clr_flag to not mute adpcm channels.
This change assumes that this line from neogeodev is incorrect:
"Flags must be manually cleared, playing a new sample on the channel won't clear it and the channel will stay silent."
In the third party source ymfm_opn used in MAME, the last phrase ("and the channel will stay silent") is not implemented.  This change attempts to match the behavior of ymfm_opn.